feat(serverless-contracts): rename getHttpLambdaHandler to getHandler
