+++ backend/orchestrator/functions/requestSyncDeployment/handler.ts
@@ -1,11 +1,11 @@
 import { ulid } from 'ulid';
 
 import { requestSyncDeployment } from '@swarmion/orchestrator-contracts';
-import { getHttpLambdaHandler } from '@swarmion/serverless-contracts';
+import { getHandler } from '@swarmion/serverless-contracts';
 
 import ServiceEventEntity from 'libs/dynamodb/models/serviceEvent';
 
-export const main = getHttpLambdaHandler(requestSyncDeployment)(async event => {
+export const main = getHandler(requestSyncDeployment)(async event => {
   const { serviceId, applicationId } = event.body;
 
   const eventId = ulid();
