+++ packages/serverless-contracts/src/contracts/apiGateway/features/index.ts
@@ -5,4 +5,3 @@ export * from './requestParameters';
 export * from './axiosRequest';
 export * from './lambdaHandler';
 export * from './fetchRequest';
-export * from './httpLambdaHandler';
