+++ user-docs/documentation/docs/how-to-guides/1-use-serverless-contracts/1-api-gateway.md
@@ -209,10 +209,10 @@ const handler = getLambdaHandler(myContract)(async event => {
 
 If you use your lambda as an ApiGateway integration, you will need to use middlewares to parse the body, validate the input and output formats and serialize the output body. You may also need to handler error cases with specific http status codes.
 
-This can be done directly by using the `getHttpLambdaHandler` function that offers all those features.
+This can be done directly by using the `getHandler` function that offers all those features.
 
 ```ts
-const handler = getHttpLambdaHandler(myContract)(async event => {
+const handler = getHandler(myContract)(async event => {
   event.pathParameters.userId; // will have type 'string'
   event.requestContext.authorizer.claims.sub; // will have type 'string' if hasAuthorize is true, otherwise will fail
